# specs/001-core-text-to-sql/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: QueryCraft — Text-to-SQL Analytics Platform API
  version: "1.0.0"
  description: |
    Phase 1 + Phase 2 API contract.
    This document is the single source of truth (Constitution Principle XII).
    Frontend types are generated from this specification.
  contact:
    name: QueryCraft Team
servers:
  - url: /api/v1
    description: API v1 base path
...
paths:
...
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [Admin]
      summary: Get admin settings
      description: Returns current application settings including LLM context cap and regenerate attempts.
      security:
        - sessionCookie: []
        - AdminKey: []
      responses:
        "200":
          description: Current admin settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettingsResponse"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin key required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAdminSettings
      tags: [Admin]
      summary: Update admin settings
      description: Updates application settings. Only provided fields are updated.
      security:
        - sessionCookie: []
        - AdminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminSettingsRequest"
      responses:
        "200":
          description: Settings updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateAdminSettingsResponse"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin key required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Invalid settings values.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─────────────────────────── Admin ───────────────────────────
  /admin/refresh-schema:
    post:
      operationId: refreshSchema
      tags: [Admin]
      summary: Force a schema introspection cache refresh
      description: |
        Clears the cached schema metadata for the source database and performs
        a fresh introspection. Returns the updated schema statistics. If the
        refreshed schema exceeds `max_schema_tokens`, returns a 422 error.
      security:
        - sessionCookie: []
        - AdminKey: []
      responses:
        "200":
          description: Schema refreshed successfully.
          content:
            application/json:
              schema:
                type: object
                required: [tables_count, columns_count, refreshed_at]
                properties:
                  tables_count:
                    type: integer
                  columns_count:
                    type: integer
                  approximate_tokens:
                    type: integer
                    description: Estimated token count of the schema context.
                  refreshed_at:
                    type: string
                    format: date-time
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Schema exceeds configured token limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
      description: |
        Server-side session backed by Redis. HttpOnly, Secure, SameSite=Strict.
        8-hour idle expiry.
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Admin endpoint authentication; configured via ADMIN_API_KEY env var.
  schemas:
...
    AdminSettingsResponse:
      type: object
      required: [llm_context_cap, max_regenerate_attempts]
      properties:
        llm_context_cap:
          type: integer
          description: Number of previous conversation turns to include as context (0-10).
        max_regenerate_attempts:
          type: integer
          description: Maximum number of regenerate attempts allowed per query including the original.
    UpdateAdminSettingsRequest:
      type: object
      required: [llm_context_cap]
      properties:
        llm_context_cap:
          type: integer
          minimum: 0
          maximum: 10
          description: Number of previous conversation turns to include as context (0-10).
        max_regenerate_attempts:
          type: integer
          minimum: 1
          description: Maximum number of regenerate attempts allowed per query including the original.
    UpdateAdminSettingsResponse:
      type: object
      required: [llm_context_cap, max_regenerate_attempts, updated_at]
      properties:
        llm_context_cap:
          type: integer
        max_regenerate_attempts:
          type: integer
        updated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error, message_key]
      properties:
        error:
          type: string
          description: Machine-readable error code.
        message_key:
          type: string
          description: i18n key for the user-facing error message. The frontend translates via i18next.
        message_params:
          type: object
          nullable: true
          additionalProperties: true
          description: Key-value map for interpolated values in the i18n template.
...